Use go/scanner to extract import path
This should be more reliable than stripping out unwanted parts of
the import line. Most notably it resolves #9

Note that I've not been able to verify this on an exhaustive set of
examples. Also note that there are still potential gotchas around
mutli line comments.. debatable whether or not those are worth
addressing (and definitely not wanting to cover in this PR, given
it has immediate benefits).