Rework parsing of imports
I mentioned in #10 that there are still potential gotchas by extracting
import info line-by-line. This attempts to address that along switching
to an even more reliable way of reading import information.

Given we're already parsing the the file for import line numbers, I
explored the idea of leaning on this information even more. As it turns
out this was viable - both comments and import path strings can be
obtained from the parsing output. This PR switches to using information
extracted solely via the parser, instead of the existing approach of
line-by-line evaluation with some information supplemented by the
parser.

Switching to the parser opened up some different possibilities. Most
notably it is now possible to deal with comments in a slightly better
way; instead of trying to pay attention but also ignore these at the
same time, the comment is handled in the same way the parser deals
with them; it is associated with the adjoining import line. This also
addresses one of the gotchas mentioned before: multi-line comments.
Assuming the multi-line comment is immediately adjoining, this is now
treated no different to single-line comments. Whilst visually these
comments may make the imports appear to be grouped separately,
semantically there is no difference. So IMO they should be treated the
same. In switching to this one other potential gotcha came to mind -
should comments be permitted that are standalone (i.e. unattached from
any import). Historically this tool has been sort of allowing for that
(certain bugs may have existed prior to switching to go/scanner). This
change allows for it, and I've covered that in the test case targeting
comment peculiarities.

Beyond that, I addressed one gotcha that wasn't discussed before: it is
entirely legal to have multiple import declarations. This is one point
where this tool likely has to become opinionated - should multiple be
permitted? Well, IMO, no (with the exception of the CGO `import "C"`
declaration). The existing implementation wouldn't take issue with this,
and when parsing line by line it would end up evaluating lines outside
of declarations (I think this is limited to comments, LPAREN and RPAREN
- which likely wouldn't trip things up after the switch to go/scanner).
The verifier now returns an error if there are multiple declarations
(`import "C"` excluded).

Finally, this addresses one final issue that #11 didn't cover - an
improperly formatted file could have unnecessary whitespace between
groups. Trimming the value of whitespace could have addressed this,
but we get a fix for free from this; the whitespace becomes irrelevant.