Prevent deadlock during path reading / results populating
Deadlock was possible before because workers were spun up prior to
the results channel being read. Using a buffered channel mitigates
this to some extent, but it's still possible to trigger on large
codebases.

Realistically this needs to be reading the results channel in
advance of the workers being spun up. This can end up getting quite
complicated, as this means spinning up and coordinating another
goroutine.

Wait groups are quite handy for this - in particular the error
handling errgroup is ideal. I've adjusted the code to spin up the
necessary goroutines in order. The workers operate on a separate
errgroup - this might seem quirky, but IMO it's the easiest way
deal with the need to close the results channel when the workers
are all done (if they were attached to the main group this would
be a logistical issue.. something would need to understand that
all workers have stopped).

NB I relocated the channels from struct level fields to variables
that are passed around. Happy to change this back, I just feel it's
a little more clear to work with, particularly given channels will
end up being closed (which would previously mean calling
`Impi.Verify()` twice would result in an error).

Resolves #13